==> 8.mix_sig2noise_train.ipynb <==
this script is for all mixing, including mixing with background noise and mixup to augment sound

import os
import glob
import soundfile as sf
import pandas as pd

# read seltab_list excel file as reference for SNR value range
seltab_list = pd.read_excel('F:/MSc Ecology & Data Science Research/Metadata/public_domainSNR.xlsx')
seltab_list.describe()

meanSNR = seltab_list['SNR'].mean()
minSNR = seltab_list['SNR'].min()
maxSNR = seltab_list['SNR'].max()

%%capture 
from audiomentations import AddBackgroundNoise, PolarityInversion

# the code is for dartmoor weak labelled data augmentation
# set up folders
input_for_high  = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_high snr_weak_call"
input_for_low   = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_low snr_weak_call"
noise_for_high  = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_high snr_weak_noise"
noise_for_low   = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_low snr_weak_noise"

output_folder_high = "F:/MSc Ecology & Data Science Research/3. train_data/1. Wilcoxon/dr_highsnr_weak/Red Fox"
output_folder_low  = "F:/MSc Ecology & Data Science Research/3. train_data/1. Wilcoxon/dr_lowsnr_weak/Red Fox"

# create the augmentation transform
# specify the range of SNR, the rest are defaults
# this function will randomly pick SNR within the specified range
transform_low = AddBackgroundNoise(
    sounds_path=noise_for_low,
    min_snr_db=minSNR,
    max_snr_db=meanSNR,
    noise_transform=PolarityInversion(),
    p=1.0
)

transform_high = AddBackgroundNoise(
    sounds_path=noise_for_high,
    min_snr_db=meanSNR,
    max_snr_db=maxSNR,
    noise_transform=PolarityInversion(),
    p=1.0
)



# create a list of wav files in input folder
sound_paths_low = glob.glob(os.path.join(input_for_low, "*.wav"))
sound_paths_high = glob.glob(os.path.join(input_for_high, "*.wav"))


# loop, augment, and save
for sound_path in sound_paths_low:
    # read sound files
    samples, sample_rate = sf.read(sound_path)
    # apply augmentation
    augmented_samples = transform_low(samples, sample_rate)
    # save back out (same filename into output_folder)
    fname = os.path.basename(sound_path)
    out_path = os.path.join(output_folder_low, fname)
    sf.write(out_path, augmented_samples, sample_rate)

# loop, augment, and save
for sound_path in sound_paths_high:
    # read sound files
    samples, sample_rate = sf.read(sound_path)
    # apply augmentation
    augmented_samples = transform_high(samples, sample_rate)
    # save back out (same filename into output_folder)
    fname = os.path.basename(sound_path)
    out_path = os.path.join(output_folder_high, fname)
    sf.write(out_path, augmented_samples, sample_rate)

%%capture 
from audiomentations import AddBackgroundNoise, PolarityInversion

# the code is for dartmoor strong labelled data augmentation
# set up folders
input_for_high  = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_high snr_strong_call"
input_for_low   = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_low snr_strong_call"
noise_for_high  = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_high snr_strong_noise"
noise_for_low   = "F:/MSc Ecology & Data Science Research/3. augment data_train/Wilcoxon/Dartmoor_low snr_strong_noise"

output_folder_high = "F:/MSc Ecology & Data Science Research/3. train_data/1. Wilcoxon/dr_highsnr_strong/Red Fox"
output_folder_low  = "F:/MSc Ecology & Data Science Research/3. train_data/1. Wilcoxon/dr_lowsnr_strong/Red Fox"

# create the augmentation transform
# specify the range of SNR, the rest are defaults
# this function will randomly pick SNR within the specified range
transform_low = AddBackgroundNoise(
    sounds_path=noise_for_low,
    min_snr_db=minSNR,
    max_snr_db=meanSNR,
    noise_transform=PolarityInversion(),
    p=1.0
)

transform_high = AddBackgroundNoise(
    sounds_path=noise_for_high,
    min_snr_db=meanSNR,
    max_snr_db=maxSNR,
    noise_transform=PolarityInversion(),
    p=1.0
)



# create a list of wav files in input folder
sound_paths_low = glob.glob(os.path.join(input_for_low, "*.wav"))
sound_paths_high = glob.glob(os.path.join(input_for_high, "*.wav"))


# loop, augment, and save
for sound_path in sound_paths_low:
    # read sound files
    samples, sample_rate = sf.read(sound_path)
    # apply augmentation
    augmented_samples = transform_low(samples, sample_rate)
    # save back out (same filename into output_folder)
    fname = os.path.basename(sound_path)
    out_path = os.path.join(output_folder_low, fname)
    sf.write(out_path, augmented_samples, sample_rate)

# loop, augment, and save
for sound_path in sound_paths_high:
    # read sound files
    samples, sample_rate = sf.read(sound_path)
    # apply augmentation
    augmented_samples = transform_high(samples, sample_rate)
    # save back out (same filename into output_folder)
    fname = os.path.basename(sound_path)
    out_path = os.path.join(output_folder_high, fname)
    sf.write(out_path, augmented_samples, sample_rate)

A path to a file/folder with sound(s), or a list of file/folder paths, must be specified. These sounds should ideally be at least as long as the input sounds to be transformed. Otherwise, the background sound will be repeated, which may sound unnatural.
What to do before mixing:
- import xeno canto metadata
- import dartmoor recording metadata
- get the length of xeno canto file 
- crop the background noise of Dartmoor recording so the length would be the same as xeno canto file
- get random dartmoor background noise
- match the background noise with xeno canto file randomly

It works, with caveat however
- if you want to get the SNR that you want, you must use the most clear signal to overlay with the noise
- if the sound already has noise, the SNR would not add linearly
- it will add noise on top of noise and you can only estimates
- so I guess for this case, only use the cleanest sound, overlay with the noise, and modify the SNR through the module
